[network + udev]: add trigger for eth interfaces
During boot USB devices may become available too late for the network script 'S40network' (init.d) to properly bring the interface(s) up. The reason is that the USB subsystem is not yet initialized at that moment. This udev rule adds a trigger add the 'eth'-interface once it becomes visible to the system.